requirements.txt
Django==1.7.1
Werkzeug==0.9.6
argparse==1.2.1
django-extensions==1.4.9
django-polymorphic==0.6
six==1.8.0
wsgiref==0.1.2
